requirements.txt
discord.py
python-dateutil
python-dotenv
requests
